+++ test_suite/test_data/__init__.py
@@ -1,0 +1,1 @@
+""" Python Bindings Unit Test for Storj (V3) """